fwoat

+++ src/core/mongo.rs
@@ -17,7 +17,7 @@ fn init_mongo() -> mongodb::error::Result<Client>
 	let client = Client::with_options(client_options)?;
 	// Ping the server to see if you can connect to the cluster
 	client.database("admin").run_command(doc! {"ping": 1}, None)?;
-	println!("Connected to server! Connection took {}s.", (sw.elapsed_ms() / 1000).to_string());
+	println!("Connected to server! Connection took {}s.", (sw.elapsed_ms() as f64 / 1000.00).to_string());
 	Ok(client)
 }
